EditItemDetails reloads itemDetails if refresh
