+++ src/components/pages/EditItemDetailsPage.js
@@ -1,76 +1,104 @@
-import React from "react";
-import { withStyles } from "@material-ui/core/styles";
-import { withRouter } from "react-router";
+import React, { useCallback, useEffect, useState } from "react";
+import { useHistory, useLocation } from "react-router";
+import { useDispatch, useSelector } from "react-redux";
+
+import _ from "lodash";
+
+import { makeStyles } from "@material-ui/core/styles";
 import Button from "@material-ui/core/Button";
-import { connect } from "react-redux";
 import Container from "@material-ui/core/Container";
+import DeleteIcon from "@material-ui/icons/Delete";
+
 import Step1 from "../AddPosting/Step1";
 import Step3 from "../AddPosting/Step3";
-import _ from "lodash";
-import DeleteIcon from "@material-ui/icons/Delete";
 import {
   updateItemDetail,
-  deletePosting,
+  deletePosting, loadItemDetail
 } from "../../redux/actions/postingActions";
 import ConfirmationDialog from "../ConfirmationDialog";
 import ImageUpload from "../ImageUpload";
+import { displayError } from "../../redux/actions/snackbarActions";
+import { LOAD_EDIT_ITEM_DETAIL_PAGE_ERROR } from "../../redux/constants/snackbarMessageTypes";
 
-const useStyles = (theme) => ({
+const useStyles = makeStyles((theme) => ({
   buttonHeader: {
     display: "flex",
     justifyContent: "space-between",
-    margin: theme.spacing(1),
+    margin: theme.spacing(1)
   },
   section: {
-    margin: theme.spacing(2, 0),
+    margin: theme.spacing(2, 0)
   },
   cancelButton: {
     color: "#AD343E",
-    borderColor: "#AD343E",
-  },
-});
-
-class EditItemDetailsPage extends React.Component {
-  constructor(props) {
-    super(props);
-    const { itemDetail } = props;
-
-    this.state = {
-      ...itemDetail,
-      errors: {
-        title: "",
-        description: "",
-        category: "",
-        condition: "",
-        quantity: "",
-      },
-      changedFields: [],
-      confirmationOpen: false,
-      confirmationType: "",
-    };
+    borderColor: "#AD343E"
   }
+}));
 
-  redirect = () => {
-    const { history } = this.props;
+const EditItemDetailsPage = () => {
+
+  const classes = useStyles();
+  const history = useHistory();
+  const location = useLocation();
+  const dispatch = useDispatch();
+  const itemDetail = useSelector(state => state.itemDetail);
+
+  const [itemDetailState, setItemDetailState] = useState({});
+  const [errors, setErrors] = useState({
+    title: "",
+    description: "",
+    category: "",
+    condition: "",
+    quantity: ""
+  });
+  const [changedFields, setChangedFields] = useState([]);
+  const [confirmationOpen, setConfirmationOpen] = useState(false);
+  const [confirmationType, setConfirmationType] = useState("");
+
+  let arrNewImgs = null;
+
+
+  const reloadItemDetails = useCallback(async () => {
+    let itemDetailsUrl = _.replace(location.pathname, "/edit", "");
+    let postingId = _.split(itemDetailsUrl, "=")[1];
+
+    let response = await dispatch(loadItemDetail(postingId));
+    if (response.type === "success") {
+      setItemDetailState(response.data);
+    } else {
+      history.goBack();
+      dispatch(displayError(LOAD_EDIT_ITEM_DETAIL_PAGE_ERROR));
+    }
+  }, [dispatch, location]);
+
+
+  useEffect(() => {
+    if (_.isEmpty(itemDetail)) {
+      reloadItemDetails();
+    } else {
+      setItemDetailState(itemDetail);
+    }
+  }, []);
+
+  const redirect = () => {
     history.goBack();
   };
 
-  isFormInvalid = () => {
-    let requiredFields = _.pick(this.state, [
+  const isFormInvalid = () => {
+    let requiredFields = _.pick(itemDetailState, [
       "title",
       "description",
       "category",
-      "condition",
+      "condition"
     ]);
     let haveEmptyFields = _.values(requiredFields).some(
       (val) => val.length === 0
     );
-    let isQuantityInvalid = this.state.quantity < 1;
+    let isQuantityInvalid = itemDetailState.quantity < 1;
     return isQuantityInvalid || haveEmptyFields;
   };
 
-  validateInput = ([key, value]) => {
-    let errors = this.state.errors;
+  const validateInput = ([key, value]) => {
     switch (key) {
       case "title":
         errors.title = value.length > 0 ? "" : "Title cannot be left blank";
@@ -91,193 +119,195 @@ class EditItemDetailsPage extends React.Component {
       default:
         break;
     }
-    this.setState({ errors: errors });
+    setErrors(errors);
   };
 
-  validateRequiredFields = () => {
+  const validateRequiredFields = () => {
     let requiredFields = _.toPairs(
-      _.pick(this.state, [
+      _.pick(itemDetailState, [
         "title",
         "description",
         "category",
         "condition",
-        "quantity",
+        "quantity"
       ])
     );
 
-    _.forEach(requiredFields, this.validateInput);
+    _.forEach(requiredFields, validateInput);
 
-    return this.isFormInvalid();
+    return isFormInvalid();
   };
 
-  updateChangedFields = (field) => {
-    if (!_.includes(this.state.changedFields, field)) {
-      this.setState({ changedFields: [...this.state.changedFields, field] });
+  const updateChangedFields = (field) => {
+    if (!_.includes(changedFields, field)) {
+      setChangedFields([...changedFields, field] );
     }
   };
 
-  handleInputChange = (e) => {
+  const handleInputChange = (e) => {
     let { name, value } = e.target;
-    this.validateInput([name, value]);
-    this.setState({ [name]: value });
-    this.updateChangedFields(name);
+    validateInput([name, value]);
+    setItemDetailState({ ...itemDetailState, [name]: value })
+    updateChangedFields(name);
   };
 
-  handleAddtoList = (e) => {
+  const handleAddtoList = (e) => {
     if (e.key === "Enter") {
       e.preventDefault();
-      this.setState({
-        [e.target.name]: [...this.state[e.target.name], e.target.value],
-      });
-      this.updateChangedFields(e.target.name);
+      setItemDetailState({...itemDetailState, [e.target.name]: [...itemDetailState[e.target.name], e.target.value]});
+      updateChangedFields(e.target.name);
       e.target.value = "";
     }
   };
 
-  handleRemoveFromList = (type, idx) => {
-    let newTags = this.state[type].filter((item, index) => index !== idx);
-    this.setState({ [type]: newTags });
-    this.updateChangedFields(type);
+  const handleRemoveFromList = (type, idx) => {
+    let newTags = itemDetailState[type].filter((item, index) => index !== idx);
+    setItemDetailState({ ...itemDetailState, [type]: newTags })
+    updateChangedFields(type);
   };
 
-  handleAddImage = (list) => {
-    this.setState({
-      images: [...this.state.images, ...list],
+  const handleAddImage = (list) => {
+    if (!arrNewImgs) {
+      arrNewImgs = itemDetailState.images;
+    }
+
+    arrNewImgs = [...arrNewImgs, ...list];
+
+    setItemDetailState({
+      ...itemDetailState,
+      images: arrNewImgs
     });
-    this.updateChangedFields("images");
+
+    updateChangedFields("images");
   };
 
-  handleConfirmationOpen = (type) => {
-    this.setState({ confirmationOpen: true, confirmationType: type });
-  };
-
-  handleSubmit = async () => {
-    if (!this.validateRequiredFields()) {
-      let details = _.pick(this.state, this.state.changedFields);
-      await this.props.updateItemDetail(this.state._id, details);
-      this.redirect();
+  const handleSubmit = async () => {
+    if (!validateRequiredFields()) {
+      let details = _.pick(itemDetailState, changedFields);
+      await dispatch(updateItemDetail(itemDetailState._id, details));
+      redirect();
     } else {
-      this.handleConfirmationOpen("alert");
+      handleConfirmationOpen("alert");
     }
   };
 
-  handleDeletePosting = async () => {
-    let { history, deletePosting } = this.props;
-    await deletePosting(this.state._id);
+  const handleDeletePosting = async () => {
+    await dispatch(deletePosting(itemDetailState._id));
     history.go(-2);
   };
 
-  handleConfirmationClose = () => {
-    this.setState({ confirmationOpen: false });
+  const handleConfirmationClose = () => {
+    setConfirmationOpen(false);
+    setConfirmationType("");
   };
 
-  selectConfirmationToDisplay = () => {
-    switch (this.state.confirmationType) {
+  const handleConfirmationOpen = (type) => {
+    setConfirmationOpen(true);
+    setConfirmationType(type);
+  };
+
+  const selectConfirmationToDisplay = () => {
+    switch (confirmationType) {
       case "delete":
         return (
           <ConfirmationDialog
-            open={this.state.confirmationOpen}
-            submitAction={this.handleDeletePosting}
+            open={confirmationOpen}
+            submitAction={handleDeletePosting}
             submitName={"Delete"}
             dialogMessage={"This action cannot be undone"}
             dialogTitle={"Are you sure you want to delete this posting?"}
-            handleClose={this.handleConfirmationClose}
+            handleClose={handleConfirmationClose}
           />
         );
       case "alert":
         return (
           <ConfirmationDialog
-            open={this.state.confirmationOpen}
-            submitAction={this.handleConfirmationClose}
+            open={confirmationOpen}
+            submitAction={handleConfirmationClose}
             submitName={"OK"}
             dialogMessage={"Please fill required the fields"}
             dialogTitle={"Information is missing from the form"}
-            handleClose={this.handleConfirmationClose}
+            handleClose={handleConfirmationClose}
             omitCancelButton={true}
           />
         );
       default:
         return (
           <ConfirmationDialog
-            open={this.state.confirmationOpen}
-            submitAction={this.redirect}
+            open={confirmationOpen}
+            submitAction={redirect}
             submitName={"OK"}
             dialogMessage={"Any changes will not be saved"}
             dialogTitle={"Are you sure you want to leave this page?"}
-            handleClose={this.handleConfirmationClose}
+            handleClose={handleConfirmationClose}
           />
         );
     }
   };
 
-  render() {
-    const { classes } = this.props;
-    return (
-      <div>
-        {this.state.confirmationOpen && this.selectConfirmationToDisplay()}
+  if (_.isEmpty(itemDetailState) || !!arrNewImgs) {
+    return null;
+  }
+
+  return (
+    <div>
+      {confirmationOpen && selectConfirmationToDisplay()}
+      <div className={classes.buttonHeader}>
+        <Button onClick={() => handleConfirmationOpen("redirect")}>
+          &lt; Back to Posting
+        </Button>
+        <Button
+          onClick={() => handleConfirmationOpen("delete")}
+          startIcon={<DeleteIcon/>}
+          className={classes.cancelButton}
+        >
+          Delete Posting
+        </Button>
+      </div>
+      <Container>
+        <div className={classes.section}>
+          <Step1
+            change={handleInputChange}
+            addTag={handleAddtoList}
+            state={{ ...itemDetailState, errors }}
+            deleteTag={handleRemoveFromList}
+          />
+        </div>
+        <div className={classes.section}>
+          <ImageUpload
+            addImage={handleAddImage}
+            images={itemDetailState.images}
+            deleteImage={handleRemoveFromList}
+          />
+        </div>
+        <div className={classes.section}>
+          <Step3
+            addRequest={handleAddtoList}
+            deleteRequest={handleRemoveFromList}
+            requests={itemDetailState.requestedItems}
+          />
+        </div>
         <div className={classes.buttonHeader}>
-          <Button onClick={() => this.handleConfirmationOpen("redirect")}>
-            &lt; Back to Posting
-          </Button>
           <Button
-            onClick={() => this.handleConfirmationOpen("delete")}
-            startIcon={<DeleteIcon />}
+            onClick={() => handleConfirmationOpen("redirect")}
+            variant="outlined"
             className={classes.cancelButton}
           >
-            Delete Posting
+            Cancel
+          </Button>
+          <Button
+            onClick={handleSubmit}
+            variant="contained"
+            color="primary"
+          >
+            Save
           </Button>
         </div>
-        <Container>
-          <div className={classes.section}>
-            <Step1
-              change={this.handleInputChange}
-              addTag={this.handleAddtoList}
-              state={this.state}
-              deleteTag={this.handleRemoveFromList}
-            />
-          </div>
-          <div className={classes.section}>
-            <ImageUpload
-              addImage={this.handleAddImage}
-              images={this.state.images}
-              deleteImage={this.handleRemoveFromList}
-            />
-          </div>
-          <div className={classes.section}>
-            <Step3
-              addRequest={this.handleAddtoList}
-              deleteRequest={this.handleRemoveFromList}
-              requests={this.state.requestedItems}
-            />
-          </div>
-          <div className={classes.buttonHeader}>
-            <Button
-              onClick={() => this.handleConfirmationOpen("redirect")}
-              variant="outlined"
-              className={classes.cancelButton}
-            >
-              Cancel
-            </Button>
-            <Button
-              onClick={this.handleSubmit}
-              variant="contained"
-              color="primary"
-            >
-              Save
-            </Button>
-          </div>
-        </Container>
-      </div>
-    );
-  }
-}
-
-const mapStateToProps = (state) => {
-  return {
-    itemDetail: state.itemDetail,
-  };
+      </Container>
+    </div>
+  );
 };
 
-export default connect(mapStateToProps, { updateItemDetail, deletePosting })(
-  withRouter(withStyles(useStyles)(EditItemDetailsPage))
-);
+
+
+export default EditItemDetailsPage;
